Move RAUC specific fixtures from conftest.py to test_rauc.py
These fixtures behave in an unexpected way as they alter the DUT's state
in a way, that we can not rely on fixtures like `shell` to ensure the
DUT's state anymore.

Since these fixtures are only used for RAUC-tests move them all into the
`test_rauc.py` so they are all in one place with the special tests that
use them.

While we are on it: Add some notes on the unusual behaviour and also add
type annotation like in the rest of this module.
