+++ conftest.py
@@ -1,5 +1,4 @@
 import contextlib
-import json
 import traceback
 from time import monotonic
 
@@ -28,53 +27,6 @@ def shell(strategy):
         pytest.exit(f"Transition into shell failed: {e}", returncode=3)
 
     return strategy.shell
-
-
-@pytest.fixture
-def default_bootstate(strategy):
-    """Set default state values as setup/teardown"""
-    strategy.transition("barebox")
-    strategy.barebox.run_check("bootchooser -a default -p default")
-
-    yield
-
-    strategy.transition("barebox")
-    strategy.barebox.run_check("bootchooser -a default -p default")
-
-
-@pytest.fixture
-def booted_slot(strategy):
-    """Returns booted slot."""
-
-    def _booted_slot():
-        strategy.transition("shell")
-        [stdout] = strategy.shell.run_check("rauc status --output-format=json", timeout=60)
-        rauc_status = json.loads(stdout)
-
-        assert "booted" in rauc_status, 'No "booted" key in rauc status json found'
-
-        return rauc_status["booted"]
-
-    yield _booted_slot
-
-
-@pytest.fixture
-def set_bootstate_in_bootloader(strategy, default_bootstate):
-    """Sets the given bootchooser parameters."""
-
-    def _set_bootstate(system0_prio, system0_attempts, system1_prio, system1_attempts):
-        strategy.transition("barebox")
-        barebox = strategy.barebox
-
-        barebox.run_check(f"state.bootstate.system0.priority={system0_prio}")
-        barebox.run_check(f"state.bootstate.system0.remaining_attempts={system0_attempts}")
-
-        barebox.run_check(f"state.bootstate.system1.priority={system1_prio}")
-        barebox.run_check(f"state.bootstate.system1.remaining_attempts={system1_attempts}")
-
-        barebox.run_check("state -s")
-
-    yield _set_bootstate
 
 
 @pytest.fixture
